feat(oscal): adds statement handling and organizes code into multiple files
